backend/requirements.txt
fastapi
uvicorn
groq
pydantic
python-dotenv
openai
pandas
jinja2
python-multipart
